refactor(host): promote `deref` utilities to top-level
Account for the fact that we will introduce Enarx-specific calls, for
some of which execution implementations will be provided
